refactor(frontend): split the oversized insurance-claims components
Addresses three Aikido findings on the insurance-claims PR:
- runAction carried an unused `claim` parameter; removed it and its call-site
  arguments.
- InsuranceClaims screen was one large function (metrics + header + states +
  table + detail + dialog); extracted InsuranceClaimsHeader and
  InsuranceClaimsStates, so it reads as composition (216 -> 130 lines).
- CreateInsuranceClaimDialog combined draft state, validation, modal behaviour
  and the whole form; extracted useInsuranceClaimDraft (draft + validate +
  build input) and InsuranceClaimFormFields, leaving the dialog as the modal
  shell (258 -> 93 lines).

Behaviour, exported props, DOM and validation are identical, so the 16 existing
tests pass unchanged; the Finance page suite stays 144/144. tsc 0, lint clean.
